==> Report/01-Introduction.ipynb <==
## Introduction

import numpy as np
import pandas as pd

Topic Modeling is an unsupervised machine learning technique, with the capability of scanning through a set of texts and clustering them, through the ability of detecting words and phrase patterns in them. This project aims to delve into these ideas by presenting models that are built to classify the type of news a specific article is presenting. These classifications being sports, entertainment, politics, tech and business. The decision to look into news articles was because we felt that it would be very interesting to see how well the models we used would word on word heavy texts. In addition, because they are formal texts, it would be standard English of a similar style from article to article. This makes the process of building a model easier because there would be little worry about slang or abbreviations, making it a good starting point. We decided that for the outcome we desired, we would be most interested in exploring bag of words and LDA. The idea is not to compete between models but to compare and contrast the strengths and weaknesses of them in this specific case. To make the bag of words useful there were two routes which were taken. One in which logistic regression was used to build a regression model and the other being naïve bayes. Below is a cleaned sample of the of the set of articles:

articles = pd.read_csv('C:/Users/Calvi/Documents/Uni/Data Science Toolbox/Assessment 2/articles_train_90_10.csv')
print('\033[1m' + "Sample of BBC Articles with Categories:")
articles.head()

This BBC news dataset made of categorised articles was retreived from [Kaggle](http://mlg.ucd.ie/datasets/bbc.html?fbclid=IwAR2BvJQxS3NxQ50e0lDOAbE4EHPR7jnpBttivWqPSyFHmfSgRpMgiKvot-4). This data contained 2225, so for building a reliable model did not seem large but we felt that because it contains full articles it seemed like there would be enough information to use. Once downloaded, the data had to be preprocessed. This was done by removing all of the unnecessary words in every article, that are not related to a specific category.
