==> Report/04-Bag of Words.ipynb <==
# Bag of Words:

import pandas as pd
import numpy as np
from math import sqrt, exp, pi
from collections import defaultdict
from sklearn.metrics import accuracy_score
from sklearn import preprocessing
from sklearn.preprocessing import StandardScaler, LabelEncoder

First must import the training and test dataset then separate the category column from the string

articles = pd.read_csv('C:/Users/Calvi/Documents/Uni/Data Science Toolbox/Assessment 2/articles_train_90_10.csv')

articles_test = pd.read_csv('C:/Users/Calvi/Documents/Uni/Data Science Toolbox/Assessment 2/articles_test_90_10.csv')

article_test_Categories = articles_test['category']

article_test_string = articles_test['string']

articles.head()

articles_wc = np.array(articles['string'])

def remove_ws(text):
    # This function turns a piece of text into a list of the words within that text
    list = []
    for i in text:
        list.append(str(i).split())
    return list

articles_ws = remove_ws(articles_wc)

def vocab():
    # This gathers all of the words mentioned in the documents and geta list of the unique ones
    vocab = []
    for i in np.array(pd.read_csv('C:/Users/Calvi/Documents/Uni/Data Science Toolbox/Assessment 2/allwords.csv')).T[0]:
        if i not in vocab:
            vocab.append(i)
    return vocab

vocab = vocab()

vocab = [x for x in vocab if not any(c.isdigit() for c in str(x))] # This removes every number mentioned

def wordtovec(text):
    # changes a list of words to a list of numbers where each one represented the number of times a certain word in vocab() is mentioned
    vector = []
    for word in vocab:
        vector.append(text.count(word))
    return vector

def wordtomat(text):
    # changes an array of lists of words to an array of lists of numbers where each row represent a text and each number in a row represent the number of times a certain word in vocab() is mentioned 
    vector = []
    counter = 0
    for article in text:
        vector.append(wordtovec(article))
        counter += 1
        
        if counter % 1000 == 0:
            print(counter)
    return vector  

bagofwords = wordtomat(articles_ws)
bagofwords_df = pd.DataFrame(bagofwords, columns = vocab)

bagofwords_df.head()

y1 = pd.DataFrame(preprocessing.LabelEncoder().fit_transform(articles['category']))

y1.columns=['Category']

bagofwords_and_cat = pd.concat([bagofwords_df,y1],axis=1)

bagofwords_and_cat.head()

## Naive Bayes Model SKlearn:

from sklearn.naive_bayes import MultinomialNB
#Model Building
classifier = MultinomialNB()
#classifier.fit(x_train,y_train)
classifier.fit(np.array(bagofwords_df),np.array(y1).flatten())

def prediction(text):
    # This gets the numbered outputs of the Naive Bayes' method and returns them to their named classes
    prediction = []
    pred = classifier.predict(text)
    for best_label in pred:
        if best_label == 0:
            prediction.append('business')
        elif best_label == 1:
            prediction.append('entertainment')
        elif best_label == 2:
            prediction.append('politics')
        elif best_label  == 3:
            prediction.append('sport')
        else:
            prediction.append('tech')
    return prediction

### Test:

bagofwords_test = wordtomat(remove_ws(np.array(article_test_string)))

pd.DataFrame(bagofwords_test).head()

pred = prediction(bagofwords_test)

print(pred)

print(len(pred))
print(article_test_Categories)

# 1 represents a correct prediction, 0 represents an incorrect prediction
array = np.zeros(len(pred))
for i in range(len(pred)):
    if(article_test_Categories[i] == pred[i]):
        array[i] = 1
    else:
        print(str(i) + ": " + article_test_Categories[i] + " and " + pred[i])
print(array)

print("Accuracy of the Naive Bayes' prediction is",sum(array)*100/len(array),"%")

## Multinomial Logistic Regression:

## Binary classification for each class

# 5 copies of the dataset

categories = np.unique(bagofwords_and_cat.iloc[:,-1])

d0 = bagofwords_and_cat.copy()
d1 = bagofwords_and_cat.copy()
d2 = bagofwords_and_cat.copy()
d3 = bagofwords_and_cat.copy()
d4 = bagofwords_and_cat.copy()

datasets = [d0,d1,d2,d3,d4]

# Turn each dataset into a binary classification

for i in range(len(categories)):
    for j in range(d0.shape[0]):
        if datasets[i].at[j,'Category'] == i:
            datasets[i].at[j,'Category'] = 1
        else:
            datasets[i].at[j,'Category'] = 0

TrainTest_datadict = {}
for i in range(len(categories)):
    X = datasets[i].iloc[:,:-1]
    X = X.T
    y = datasets[i].iloc[:,-1]
    y = np.array(y)
    TrainTest_datadict['D'+str(i)] = [X,y]

iterations = 100 # large number
alpha = 0.01
m = d0.shape[1]
cost_values = []
trained_parameters = []

for trainsets in range(len(categories)):
    X = TrainTest_datadict['D'+str(trainsets)][0]
    y = TrainTest_datadict['D'+str(trainsets)][1]
    weights = np.full((1,X.shape[0]),0.01)
    bias = 0
    costfunc_values = []
    k = 0
    print('Training for dataset '+str(trainsets))
    for i in range(1,iterations+1):
        # logistic function
        z = np.dot(weights,X) + bias
        hypothesis = 1/(1+np.exp(-z))
        
        # Cost function
        j = 1/m *(-1*(np.sum(y*np.log(hypothesis) + (1-y)*np.log(1-hypothesis))))
        costfunc_values.append(j)
        k+=1
        
        # Gradient decent
        
        dw = 1/m * np.dot(hypothesis-y,X.T)
        db = 1/m * np.sum(hypothesis-y)
        weights = weights - alpha*dw
        bias = bias - alpha*db
        
        # stop training
        if i%2000 == 0:
            print('running @',j)
        if i%2 == 0:
            if abs(j-costfunc_values[-2])<0.000001:
                if abs(j-costfunc_values[-3])<0.000001:
                    break
        
        cost_values.append(costfunc_values)
        trained_parameters.append([weights,bias])
        print('iteration number:',k)

### Test:

test = pd.DataFrame(bagofwords_test)

def prediction(trained_parameters,test):
    inputs = test.T
    result = []
    prediction = []
    print(inputs.shape[1])
    for i in range(inputs.shape[1]):
        probabilities = []
        for j in range(5):
            weights = trained_parameters[j][0]
            bias = trained_parameters[j][1]
            z = np.dot(weights,(inputs.iloc[:,i]).T)+bias
            hypothesis = 1/(1+np.exp(-z))
            probabilities.append(hypothesis)
        predict = probabilities.index(max(probabilities))
        result.append(predict)
    for best_label in result:
        if best_label == 0:
            prediction.append('business')
        elif best_label == 1:
            prediction.append('entertainment')
        elif best_label == 2:
            prediction.append('politics')
        elif best_label  == 3:
            prediction.append('sport')
        else:
            prediction.append('tech')
    return prediction

pred = prediction(trained_parameters,test)

print(pred)

There seems to be an error with the multinomial logisitic regression attempt. It only outputs business.
